product-ms/src/main/java/com/nexeum/productms/controller/WebController.java
package com.nexeum.productms.controller;

import com.nexeum.productms.service.ProductService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1/web/product")
public class WebController {
    private final ProductService productService;

    public WebController(ProductService productService) {
        this.productService = productService;
    }

    @GetMapping("/all-products")
    public ResponseEntity<Object> getAllProducts() {
        return productService.getAllProducts();
    }
}
